Cleanup the Platform trait
This commit provides default implementations for some methods of the
Platform trait which were shared by multiple platforms. It helps tidying
up our existing platform implementations, but does not introduce any
semantic changes.
