+++ src/platform/premierp550.rs
@@ -1,13 +1,12 @@
 //! SiFive HiFive P550 board
 
+use core::fmt;
 use core::fmt::Write;
-use core::{fmt, hint};
 
 use log::Level;
 use spin::Mutex;
 
-use crate::arch::{read_custom_csr, write_custom_csr, Arch, Architecture};
-use crate::config::{TARGET_FIRMWARE_ADDRESS, TARGET_START_ADDRESS};
+use crate::arch::{read_custom_csr, write_custom_csr};
 use crate::device::clint::{VirtClint, CLINT_SIZE};
 use crate::device::VirtDevice;
 use crate::driver::clint::ClintDriver;
@@ -16,8 +15,6 @@ use crate::Platform;
 
 // —————————————————————————— Platform Parameters ——————————————————————————— //
 
-const MIRALIS_START_ADDR: usize = TARGET_START_ADDRESS;
-const FIRMWARE_START_ADDR: usize = TARGET_FIRMWARE_ADDRESS;
 const CLINT_BASE: usize = 0x0000_0200_0000;
 
 // ———————————————————————————— Platform Devices ———————————————————————————— //
@@ -68,32 +65,6 @@ impl Platform for PremierP550Platform {
         let mut writer = WRITER.lock();
         writer.write_fmt(args).unwrap();
         writer.write_str("\r").unwrap();
-    }
-
-    fn exit_success() -> ! {
-        loop {
-            Arch::wfi();
-            hint::spin_loop();
-        }
-    }
-
-    fn exit_failure() -> ! {
-        loop {
-            Arch::wfi();
-            hint::spin_loop();
-        }
-    }
-
-    fn load_firmware() -> usize {
-        FIRMWARE_START_ADDR
-    }
-
-    fn get_miralis_start() -> usize {
-        MIRALIS_START_ADDR
-    }
-
-    fn get_max_valid_address() -> usize {
-        usize::MAX
     }
 
     fn get_virtual_devices() -> &'static [VirtDevice] {
